+++ CHANGELOG.md
@@ -40,6 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#69](https://github.com/alleslabs/celatone-frontend/pull/69) Add execute table in contract details page
+- [#63](https://github.com/alleslabs/celatone-frontend/pull/63) Add code id explorer link and code table row navigation
+- [#67](https://github.com/alleslabs/celatone-frontend/pull/67) Add Public Codes shortcut to sidebar and add Quick Actions section
 - [#66](https://github.com/alleslabs/celatone-frontend/pull/66) Add code details data loader including code info and contract instances
 - [#60](https://github.com/alleslabs/celatone-frontend/pull/60) Add navigation to contract row
 - [#47](https://github.com/alleslabs/celatone-frontend/pull/47) Wireup init msg in contract details page